Run Modular tests
The Fuchsia recipe now has a "tests" property that takes the path of
a test runner JSON config file. Setting this property will cause the
tests to run in QEMU after every build.

However, I still need to implement checking for pass/fail in the actual
tests. As of this CL, we're only seeing if the run_test script will fail
to complete. When testing locally, I have seen it fail on a TCP hangup
from the test_runner inside Fuchsia, so that's something to look out
for.

Change-Id: I3e43745372d48466aab29d4af0ba76a4024483a3