Switch Ledger to Fuchsia recipe with tests
This is going to be a slower ninja build step for the moment, since
the "tests" property causes the "boot_tests_modular" module to be pulled
into the build, which includes "default", which is everything.

I'll need to write a few CLs to fix that, but in the meantime something
is better than nothing.

Change-Id: Id54e4b32ec152fb7669f9a6b209fcdfb64f4d1c9