Use Fuchsia primary builders for Mondrian, SysUI
There was a confusing failure because the Mondrian and SysUI commit
queues were using the "fuchsia" mixin, which runs all of the Fuchsia
tests and assumes a full Fuchsia build. However, the build was being
restricted, so test binaries were not present, causing CQ failure.

Since the "mondrian" and "sysui" build modules are not in widespread
use for development, they might not be up-to-date and valid for CQ use.
It's simpler and safer (although slower) to just use the primary Fuchsia
builders on CQ for these repos, at least until we have a better
dependency story in Fuchsia overall.

This change also adds a comment to the "fuchsia" mixin, explaining its
purpose.

Change-Id: I3119ed88a03127dd6ba2ec7f2c18b8a87a8d3b79