Restore cobalt-client Luci scheduler.
In https://fuchsia-review.googlesource.com/#/c/infra/config/+/65927/
the cobalt-client CI builders were removed. The patch description
says because they "have been superceded by Garnet builders". But
this is not true because the CI that was removed runs a slightly
different configuration of the tests than the one in Garnet.
The one that was removed uses the network to communicate
with the real Cobalt servers in the cloud. The version in
Garnet does not do this.

Here we restore CI for cobalt-client that uses the network.
This makes the test is not completely hermetic. But the benefit
is that we get to test the full stack of the Cobalt client.
The non-hermeticness is not a big problem because only
people on the Cobalt team will be paying attention to this
particular CI and we can understand if the tests fail
because of problems with the Cobalt server.

In this CL we also remove the CQ builders for cobalt_client.
These are not necessary anymore as they have been superceded by
Garnet builders.

Change-Id: Icdc602da46ace9ffe308d20e65422109a597a6c0