Enable host_asan on Garnet and Peridot builders
The effect on bot cycle times is small enough not to warrant
separate bots for ASan testing of the build's use of host tools.

The effect on topaz-x64 is enough to give pause, so not enabling
there yet--meaning no ASan testing of host-side Dart on bots.
Due to TO-660, the effect on topaz-arm64 (host-side Dart when
targetting arm64) is so huge it's not even viable for FYI bots.

IN-132 #comment Done for layers below Topaz

Change-Id: I303ad89827c69ceff13eb8e3de63f05b00ca352a